refactor: move declaration to separate .h file
- function declarations moved to .h to avoid broken includes in cgo code
- extern was kept at .c since this information only needs to be know at
  link time
