+++ pkg/quic/lsquic/adapter.c
@@ -1,56 +1,22 @@
-#include <errno.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdlib.h>
 
+#include "adapter.h"
 #include "logger.h"
-#include "lsquic.h"
+#include "server.h"
 
-struct lsquic_conn_ctx {
-    void* adapter_ctx;
-};
+#ifndef ADAPTER
+#define ADAPTER
 
-struct lsquic_stream_ctx {
-    char* send_buffer;
-    size_t send_buffer_size;
-    off_t send_buffer_off;
-    void* adapter_ctx;
-};
-
-/* Adapter Callbacks */
 extern lsquic_conn_ctx_t* adapterOnNewConnection(lsquic_conn_t* conn, void* stream_if_ctx);
 extern void adapterOnClosedConnection(lsquic_conn_t* conn);
-extern lsquic_stream_ctx_t* adapterOnNewStream(lsquic_stream_t* stream, void* adapter_ctx);
-extern void adapterOnRead(lsquic_stream_t* stream, char* buf, size_t buf_size, void* adapter_ctx);
-extern void adapterOnWrite(lsquic_stream_t* stream, void* adapter_ctx);
+extern lsquic_stream_ctx_t* adapterOnNewStream(lsquic_stream_t* stream, void* stream_ctx);
+extern void adapterOnRead(lsquic_stream_t* stream, char* buf, size_t buf_size, lsquic_stream_ctx_t* stream_ctx);
+extern void adapterOnWrite(lsquic_stream_t* stream, lsquic_stream_ctx_t* stream_ctx);
 extern void adapterOnClose(lsquic_stream_t* stream, lsquic_stream_ctx_t* stream_ctx);
 
-/**/
-
-/* LSQUIC Callbacks */
-
-/*
- * Callback to process the event of a new connection to the server
- * */
-extern lsquic_conn_ctx_t* server_on_new_connection(void* stream_if_ctx,
-    struct lsquic_conn* conn);
-
-/*
- * Callback to process the event of a closed connection to the server
- * */
-extern void server_on_closed_connection(lsquic_conn_t* conn);
-
-extern lsquic_stream_ctx_t* server_on_new_stream(void* stream_if_ctx,
-    struct lsquic_stream* stream);
-
-extern void server_on_read(struct lsquic_stream* stream,
-    lsquic_stream_ctx_t* stream_ctx);
-
-extern void server_on_write(struct lsquic_stream* stream,
-    lsquic_stream_ctx_t* stream_ctx);
-
-extern void server_on_close(struct lsquic_stream* stream,
-    lsquic_stream_ctx_t* stream_ctx);
+/* Adapter Callbacks */
 
 static const struct lsquic_stream_if stream_interface = {
     .on_new_conn = server_on_new_connection,
@@ -61,19 +27,15 @@ static const struct lsquic_stream_if stream_interface = {
     .on_close = server_on_close,
 };
 
-/* stream methods */
-
-void print_conn_info(const lsquic_conn_t* conn)
+bool lsquic_new_server(
+    Server* server,
+    const char* keylog,
+    void* stream_if_ctx)
 {
-    const char* cipher;
-
-    cipher = lsquic_conn_crypto_cipher(conn);
-
-    Log("Connection info: version: %u; cipher: %s; key size: %d, alg key size: "
-        "%d",
-        lsquic_conn_quic_version(conn), cipher ? cipher : "<null>",
-        lsquic_conn_crypto_keysize(conn), lsquic_conn_crypto_alg_keysize(conn));
+    return new_server(server, keylog, &stream_interface, stream_if_ctx);
 }
+
+/* stream methods */
 
 lsquic_conn_ctx_t* server_on_new_connection(void* stream_if_ctx,
     struct lsquic_conn* conn)
@@ -98,31 +60,9 @@ lsquic_stream_ctx_t* server_on_new_stream(void* stream_if_ctx,
 void server_on_read(struct lsquic_stream* stream,
     lsquic_stream_ctx_t* stream_ctx)
 {
-    struct lsquic_stream_ctx* const stream_data = (void*)stream_ctx;
     ssize_t num_read;
-    unsigned char buf[0x400];
-
-    if (stream_ctx == NULL) {
-        Log("in %s: received NULL context", __func__);
-        lsquic_stream_close(stream);
-        return;
-    }
-
-    lsquic_stream_id_t id = lsquic_stream_id(stream);
-    Log("Trying to read from Stream with id: %d", id);
-    num_read = lsquic_stream_read(stream, buf, sizeof(buf));
-
-    if (num_read < 0) {
-        /* This should not happen */
-        Log("error reading from stream (errno: %d) -- abort connection", errno);
-        lsquic_conn_abort(lsquic_stream_conn(stream));
-    }
-
-    Log("read %ld bytes", num_read);
-    lsquic_stream_wantread(stream, false);
-    lsquic_stream_wantwrite(stream, true);
-    // TODO: callback to treat data
-    // lsquic_stream_wantwrite(stream, 1);
+    char buf[0x400];
+    adapterOnRead(stream, buf, sizeof(buf), stream_ctx);
     return;
 }
 
@@ -131,14 +71,8 @@ void server_on_write(struct lsquic_stream* stream,
 {
     ssize_t num_written;
     lsquic_stream_id_t id = lsquic_stream_id(stream);
-    num_written = lsquic_stream_write(stream, stream_ctx->send_buffer, stream_ctx->send_buffer_size);
-    lsquic_stream_flush(stream);
-    if (num_written < 0) {
-        Log("lsquic_stream_write() returned %ld, abort connection",
-            (long)num_written);
-        lsquic_conn_abort(lsquic_stream_conn(stream));
-    }
-    Log("All data was written back, stopping stream");
+    Log("trying to write to stream with id: #%u", id);
+    adapterOnWrite(stream, stream_ctx);
 }
 
 void server_on_close(struct lsquic_stream* stream,
@@ -148,3 +82,4 @@ void server_on_close(struct lsquic_stream* stream,
     Log("%s called, closing stream %u", __func__, id);
     adapterOnClose(stream, stream_ctx);
 }
+#endif
